% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GICS-package.R
\docType{package}
\name{GICS-package}
\alias{GICS}
\alias{GICS-package}
\title{GICS: Global Industry Classification Standard dataset}
\description{
Global Industry Classification Standard conveniently packaged for consumption in R.
}
\author{
\strong{Maintainer}: Olivier Bautheac \email{o.bautheac@gmail.com}

}
\keyword{internal}
