% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GICS-package.R
\docType{package}
\name{GICS-package}
\alias{GICS}
\alias{GICS-package}
\title{GICS: Global Industry Classification Standard dataset}
\description{
Global Industry Classification Standard conveniently packaged for consumption in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bautheac.github.io/GICS}
  \item Report bugs at \url{https://github.com/bautheac/GICS/issues}
}

}
\author{
\strong{Maintainer}: Olivier Bauthéac \email{o.bautheac@gmail.com}

}
\keyword{internal}
